requirements.txt
aiosqlite
bcrypt
cffi
fastapi
h11
h2
hpack
uvicorn
hyperframe
iso8601
passlib
priority
pycparser
pydantic
PyJWT
PyPika
python-multipart
pytz
six
starlette
toml
tortoise-orm
typing-extensions
wsproto